# docs/openapi/veil.openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Veil",
  "version": "0.1.1",
  "description": "Agent credential broker. Use injects the secret at the edge. Responses never include vault material. CLI, MCP, and generated SDKs are projections of this contract."
 },
 "servers": [
  {
   "url": "https://veil.nyc"
  }
 ],
...
 "security": [
  {
   "bearerAuth": []
  }
 ],
 "paths": {
...
  "/v1/members/{id}": {
   "delete": {
    "operationId": "removeMember",
    "tags": [
     "Human"
    ],
    "summary": "Offboard a member. Owner-only \u2014 drops the Keto member tuple and the humans row; the removed member's token resolves nothing from that call on. Owners cannot be removed this way (demote first). Not MCP.",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Kratos identity id of the member"
     }
    ],
    "responses": {
     "200": {
      "description": "Member removed",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "removed": {
           "type": "boolean"
          }
         }
        }
       }
      }
     },
     "401": {
      "description": "missing or invalid human token"
     },
     "403": {
      "description": "caller is a member, not an owner"
     },
     "404": {
      "description": "not a member of this org"
     }
    }
   }
  },
  "/v1/members/{id}/owner": {
   "post": {
    "operationId": "promoteOwner",
    "tags": [
     "Human"
    ],
    "summary": "Grant an existing member the owners tuple. Owner-only. Not MCP.",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Kratos identity id of the member"
     }
    ],
    "responses": {
     "200": {
      "description": "Member promoted",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "owner": {
           "type": "boolean"
          }
         }
        }
       }
      }
     },
     "401": {
      "description": "missing or invalid human token"
     },
     "403": {
      "description": "caller is a member, not an owner"
     },
     "404": {
      "description": "not a member of this org"
     }
    }
   },
   "delete": {
    "operationId": "demoteOwner",
    "tags": [
     "Human"
    ],
    "summary": "Strip the owners tuple from a co-owner. Owner-only. The last owner cannot be demoted \u2014 the org would have no administrator. Not MCP.",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      },
      "description": "Kratos identity id of the owner"
     }
    ],
    "responses": {
     "200": {
      "description": "Owner demoted",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "owner": {
           "type": "boolean"
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "last-owner demotion refused"
     },
     "401": {
      "description": "missing or invalid human token"
     },
     "403": {
      "description": "caller is a member, not an owner"
     },
     "404": {
      "description": "not an owner of this org"
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "bearerAuth": {
    "type": "http",
    "scheme": "bearer",
    "bearerFormat": "JWT",
    "description": "Hydra JWT. Identity is the bound agent or a Keto member human. The host is not identity."
   }
  },
...
 }
}